% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture_fracts_from_ancestry_tracts.R
\name{admixture_fracts_from_ancestry_tracts}
\alias{admixture_fracts_from_ancestry_tracts}
\title{calculate the admixture fraction of indivdiduals from ancestry tracts}
\usage{
admixture_fracts_from_ancestry_tracts(AT)
}
\arguments{
\item{AT}{the tibble of ancestry tracts like what comes out of \code{\link[=indiv_ancestry_tracts]{indiv_ancestry_tracts()}}.}
}
\description{
Super simple.  Just counting up the length in different ancestries, summing
the two diploid genotypes together.  This should extend to multiple ancestries
directly
}
