% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_genome_position.R
\name{cumulative_genome_position}
\alias{cumulative_genome_position}
\title{Add a column called cPOS that is the cumulative genome position}
\usage{
cumulative_genome_position(X, Cumul)
}
\arguments{
\item{X}{a tibble that has columns of \code{chrom} and \code{pos}}

\item{Cumul}{a tibble that has, at a minimum, the columns \code{chrom}, and \code{cumul_start}.
It will often be easiest to pass in your genome chromosome cumulative tibble
like \code{\link{mykiss_chroms}}}
}
\value{
Returns a tibble like X but with an additional column cPOS that has the
cumulative position.
}
\description{
SLiM does everythign in terms of cumulative genome positions (i.e., there is
just one chromosome that happens to have locations where the recombination
fraction if 1/2).  So, if we have a genome with coordinates of chrom, pos,
we might want to get the cPOS for each.  Doing so can be done efficiently
by precomputing the cumulative start point of each chromosome like we
have in the package data object \code{\link{mykiss_chroms}}.
}
\examples{
cumulative_genome_position(example_spp_diag_markers, mykiss_chroms)

}
