% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyclone_creek_var_freqs}
\alias{cyclone_creek_var_freqs}
\title{Example frequencies of alleles in two ancestries}
\format{
A tibble
}
\usage{
cyclone_creek_var_freqs
}
\description{
These were estimated from cyclone creek and are mapped to the
mykiss genome of \code{mykiss_chroms}
}
\keyword{datasets}
