% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_indiv_segs}
\alias{example_indiv_segs}
\title{Example indiv_segs.}
\format{
See the documentation for \code{\link[=indiv_ancestry_tracts]{indiv_ancestry_tracts()}}.
}
\usage{
example_indiv_segs
}
\description{
This is an example of the sort of output that comes out of \code{\link[=indiv_ancestry_tracts]{indiv_ancestry_tracts()}}.
This particular data set includes 10 individuals simulated when
developing this package.
}
\keyword{datasets}
