% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_spp_diag_markers}
\alias{example_spp_diag_markers}
\title{Example data set of positions of species-diagnostic markers}
\format{
A tibble with columns:
\itemize{
\item \code{diag_spp}: 1 = WCT, 2 = RBT, 3 = YCT
\item \code{chrom}: the chromosome in the mykiss genome
\item \code{pos}: the position of the marker on chromosome chrom.
}
}
\usage{
example_spp_diag_markers
}
\description{
These were pulled from the Cyclone Creek data set.
}
\keyword{datasets}
