% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/founder_node_ids_and_pops.R
\name{founder_node_ids_and_pops}
\alias{founder_node_ids_and_pops}
\title{get the ids and pops of founder nodes}
\usage{
founder_node_ids_and_pops(NI, FY)
}
\arguments{
\item{NI}{a list with tibbles of tskit nodes and individuals like that
returned by \code{ts_nodes_and_inds()}.}

\item{FY}{the year (in tskit time) in which individuals are
designated as founders.}
}
\description{
The "Founders" are defined according to the time they
existed.  You give a year (in tskit time) and this returns
a list with two elements: $nodes---a vector of ids of the
founder nodes, and $pops---a vector of the pops that the
founder nodes belonged to.
}
