% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_focal_nodes.R
\name{get_focal_nodes}
\alias{get_focal_nodes}
\title{Get list of nodes at which to determine ancestry segments}
\usage{
get_focal_nodes(NI, Focal)
}
\arguments{
\item{NI}{The list of nodes and inds tibbles like that returned from
\code{ts_nodes_and_inds()}.}

\item{Focal}{a tibble with two columns: \code{pop} and \code{time}.  Each row is a combination
of a pop in the simulation and a cohort year that you want included amongst the
focal nodes.}
}
\value{
Returns a list of vectors.  Each vector is a vector of node IDs that
are from the same cohort.
}
\description{
When we link ancestors to the focal nodes we have to be careful!
If we include in the sample, any individuals that other individuals in the
sample descended from, then these "sampled ancestors" are what gets returned
rather than the actual earlier "founders."
So, we actually need to link_ancestors separately for each cohort of the
samples with the ancestors being the same ancestors every time.
}
\details{
This function prepares a list of focal nodes, with each element being
focal nodes from a particular cohort.  This is what the \code{link_ancestors()}
function expects as input.
}
