% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_ancestry_tracts.R
\name{indiv_ancestry_tracts}
\alias{indiv_ancestry_tracts}
\title{get ancestry tracts of individuals}
\usage{
indiv_ancestry_tracts(A)
}
\arguments{
\item{A}{a tibble of ancestral segments like that produced by \code{ancestral_segs()}.}
}
\description{
This takes the output of \code{ancestral_segs()} and returns a tibble of the
ancestry doses of the individuals themselves.  To do this, the function
relies on \code{bedtools} which must be on your path.
}
