% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_relationships.R
\name{pairwise_relationships}
\alias{pairwise_relationships}
\title{Return pairwise relationships between the samples given the population pedigree}
\usage{
pairwise_relationships(IT, sample_years, num_gen)
}
\arguments{
\item{IT}{The \code{indiv_tib} component of the return list from
\code{ts_nodes_and_inds()}.  Or any tibble with at least with columns
\code{ind_time}, \code{sex}, \code{ped_id}, \code{ped_p1}, and \code{ped_p2}.}

\item{sample_years}{A vector of years in TSKIT time from which all the
individuals in the pedigree will be deemed the samples.}

\item{num_gen}{number of generations back from the sampled individuals to return in each
individual's ancestor vector. 0 = self, 1 =  back to and including the parents,
2 = back to and including the grandparents, and so on.}
}
\description{
This uses some functions from CKMRpop to deliver relationships.
}
\details{
Note that for convenience, we discard the individual with ped_id 0,
which in-effect means that ped_id 0 just becomes another unobserved
founder of any other individuals.  Unless 0 is within num_gen generations
of any of your samples, this should have no effect at all.
}
