% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_diagnostic_markers.R
\name{place_diagnostic_markers}
\alias{place_diagnostic_markers}
\title{Place diagnostic marker alleles within individauls according to their ancestry haplotypes}
\usage{
place_diagnostic_markers(Markers, IndivSegs)
}
\arguments{
\item{Markers}{a tibble that \emph{must} have columns:
\itemize{
\item \code{diag_spp_idx}, which gives a 1-based
index of the species for which the \code{1} allele is fixed at the marker (with all
other species being fixed for the \code{0} allele).
\item \code{cPOS}, which are positions in the cumulative base pairs
that SLiM uses.
\item \code{chrom}: the name of the chromosome (in the actual genome being modeled)
\item \code{pos}: the position in base pairs on chromosome \code{chrom}.
Each row of this tibble must specify a unique species-diagnostic marker, and
all of the species-diagnostic markers you wish to place into the simulated
individuals should be included in this tibble.
}}

\item{IndivSegs}{A tibble of individual segments like that which is returned
by the function \code{\link[=indiv_ancestry_tracts]{indiv_ancestry_tracts()}}. Though it can have arbitrary columns,
it must include at least the following:
\itemize{
\item \code{ped_id}: The pedigree id of the individual
\item \code{left}: the left endpoint of the interval
\item \code{right}: the right endpoint of the interval
\item \code{anc1}: the 1-based index of the ancestry of haplotype 1 within the interval.
\item \code{anc2}: the 1-based index of the ancestry of haplotype 2 within the interval.
}}
}
\value{
This returns a tibble with the columns \code{diag_spp_idx}, \code{chrom}, \code{pos},
\code{ped_id}, and \code{geno}.  This is almost what is needed to hand the data off to
the \code{ancestral_n_segments()} function in the MixedUpParents package. You just
have to change some column names and denote diag_spp by A, B, C.
}
\description{
Once individuals have been simulated and you have the individual segments
within them, you can easily simulate the alleles of the species-diagnostic
markers. That is what this function does.
}
\details{
For this to work, you have to have both bedtools and awk on your PATH.
}
\examples{
# Prepare Markers. Name the columns right, and get cumulative cPOS in there.
M <- example_spp_diag_markers \%>\%
  dplyr::filter(diag_spp < 3) \%>\%
  dplyr::rename(diag_spp_idx = diag_spp) \%>\%
  cumulative_genome_position(Cumul = mykiss_chroms)

# Now, also get some example indiv-segs. We have some as a package
# data object.
IS <- example_indiv_segs

# Then run the function:
place_diagnostic_markers(M, IS)
}
