% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ancestry_copy_number.R
\name{plot_ancestry_copy_number}
\alias{plot_ancestry_copy_number}
\title{Plot the ancestry copy number along segments in a group of individuals}
\usage{
plot_ancestry_copy_number(IndSegs, chrom_ends = NULL)
}
\arguments{
\item{IndSegs}{the tibble that comes out of \code{\link[=indiv_ancestry_tracts]{indiv_ancestry_tracts()}}.}

\item{chrom_ends}{if desired, a tibble with column \code{chrom} giving the
chromosome name and column \code{end} giving the cumulative base pair
position of the end of the chromosome.  Makes the plot return horizontal
lines at the chromosome boundaries.}
}
\value{
returns a ggplot object
}
\description{
More documentation needed.  Basically just plots what comes out of
\code{\link[=indiv_ancestry_tracts]{indiv_ancestry_tracts()}}.  Currently assumes just two ancestries.
Plots individuals in order of admixture fraction.
}
