% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ancestry_of_genomes.R
\name{plot_ancestry_of_genomes}
\alias{plot_ancestry_of_genomes}
\title{Plot genome segments in focal individuals colored by ancestral population}
\usage{
plot_ancestry_of_genomes(Segs)
}
\arguments{
\item{Segs}{a tibble of segments like what comes out of \code{ancestral_segs()}}
}
\value{
a ggplot object
}
\description{
This just wraps up some ggplot.  Note that this plots the ancestry of each
of the two distinct genomes within an individual, but it does not plot the
two genomes within an individual necessarily adjacent to one another. For that
see \code{plot_ancestry_copy_number()}
}
