% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_founder_genos.R
\name{simulate_founder_genos}
\alias{simulate_founder_genos}
\title{Simulate genotypes for founders from allele frequencies}
\usage{
simulate_founder_genos(Freqs, N, pop_names)
}
\arguments{
\item{Freqs}{A tibble with a column \code{POS} which are positions in the cumulative base pairs
that SLiM uses, and then columns \code{afreq_1}, \code{afreq_2}, \code{afreq_3}, etc.  These are the
alternate allele freqs in populations 1, 2, and 3, and so on. Note that the names
of those columns don't really matter---what matters is their position. The first
must be the freqs for population 1, the second for population 2, and so forth.}

\item{N}{A vector of number of diploids to simulate for each of the populations.}

\item{pop_names}{A vector of the same length as N, with short names for pops.
No whitespace in the names, please!}
}
\value{
Returns a tibble with first column POS and subsequent columns
named pop_name + 0-based index.  Includes genotypes from all the
individuals in all the pops.
}
\description{
This takes in allele frequencies for different populations and the number
of individuals in those populations, and simulates genotypes for them into
a tibble.
}
\examples{
# make allele freqs
freq <- tibble::tibble(
  POS = seq(10, 90, by = 10),
  af1 = rep(0.1, 9),
  af2 = rep(0.9, 9),
  af3 = rep(0.5, 9)
)

# number of diploids to simulate
n <- c(3, 5, 7)

# names of populations
pn <- c("a", "b", "c")

# run the function
simulate_founder_genos(freq, n, pn)
}
