% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slim_sim_dataset_example}
\alias{slim_sim_dataset_example}
\title{Example output of \code{slim_sim_a_dataset()}}
\format{
An object of class \code{list} of length 8.
}
\usage{
slim_sim_dataset_example
}
\description{
This is just some output from running the example of
\code{slim_sim_a_dataset()} so that it can be used in developing
and illustrating the function \code{tweak_slim_simmed_data.R}.
}
\keyword{datasets}
