% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_vcf2tib.R
\name{slim_vcf2tib}
\alias{slim_vcf2tib}
\title{Read a single SLiM-formatted VCF into a tibble of genotypes}
\usage{
slim_vcf2tib(path, long = TRUE)
}
\arguments{
\item{path}{The path to the VCF file.}

\item{long}{Logical. If true (the default) this returns the long format
version.  Otherwise it returns the wide format.}
}
\description{
By default this makes it long format, but if you want to keep it wide
format, you can request long = FALSE. Wide format could be better
if you need to make a wide-format data set out of it, or for space
purposes.
}
\details{
MOre
}
