% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_nodes_and_inds.R
\name{ts_nodes_and_inds}
\alias{ts_nodes_and_inds}
\title{return a tibble of the nodes and indviduals and meta data from a tree sequence}
\usage{
ts_nodes_and_inds(ts)
}
\arguments{
\item{ts}{the tree sequence object as you would access it from the R environment,
for example, \code{py$ts} or \code{py$tree}, etc.}
}
\description{
The way this works is a little non-standard.  It writes the input, \code{ts} into
the global environment under the name xx__TTKKSS where python can act upon it.
Then a bunch of other stuff gets written into the python environment which is
then pulled back into a tibble.  At the end, we try to remove all the \code{xx__}
variables that we put into different global environments.
}
