% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak_slim_simmed_data.R
\name{tweak_slim_simmed_data}
\alias{tweak_slim_simmed_data}
\title{Subsample SNPs or individuals and simulated genotyping error}
\usage{
tweak_slim_simmed_data(
  X,
  NV = NA,
  ND = NA,
  S = NA,
  RV = NA,
  RD = NA,
  EV = NA,
  ED = NA
)
}
\arguments{
\item{X}{the return object of \code{slim_sim_a_dataset()}.}

\item{NV}{the number of variable markers to randomly downsample to, retaining
\emph{the same markers across all individuals} and completely discarding
the rest. If \code{NV} is greater than the total number of variable markers
a warning is issued and the full set of markers is returned.}

\item{ND}{the number of diagnostic markers to randomly downsample to, retaining
\emph{the same markers across all individuals} and completely discarding
the rest. If \code{ND} is greater than the total number of diagnostic markers
a warning is issued and the full set of markers is returned.}

\item{S}{the fraction of sampled individuals to be randomly
\emph{retained}.  So, if you want to have sampled 40\% of the population
then you would set S = 0.4.}

\item{RV}{the missing data retainment data rate for variable markers.  This is
applied randomly to each marker in each individual, after subsampling
to NV markers.  A fraction 1 - RV is expected to be missing.}

\item{RD}{the missing data retainment rate for diagnostic markers.  This is
applied randomly to each marker in each individual, after subsampling
to ND markers.  A fraction 1 - RD is expected to be missing after
this.}

\item{EV}{the probability that any non-missing variable marker genotype in an individual
has a genotyping error.  For a genotype that is simulated to have
a genotyping error, if the true genotype is a 1 we choose either 0 or 1, each
with probability of 1/2. If the true genotype is a 0 or a 2, the
genotyping error results in a 1.  (Genotyping error to the opposite
homozygote does not occur).}

\item{ED}{the probability that any non-missing diagnostic marker genotypes in an individual
has a genotyping error.  For a genotype that is simulated to have
a genotyping error, if the true genotype is a 1 we choose either 0 or 1, each
with probability of 1/2. If the true genotype is a 0 or a 2, the
genotyping error results in a 1.  (Genotyping error to the opposite
homozygote does not occur).}
}
\value{
This returns a list with the downsampled and genotyping-errored
versions of the data set.
}
\description{
Once you have simulated a SLIM dataset with \code{slim_sim_a_dataset()},
you can use this function to do the following:
\itemize{
\item Downsample to a consistent subsample of some number NV variable
markers and/or some number ND diagnostic ones.
\item Create missing data at the remaining markers at a simple rate
RV or RD across variable or diagnostic markers, respectively.
\item Randomly toss each individual in the sample from the sample
at a single rate S across all individuals, pops and times.
\item Downsample the number of individuals to a specific number or
fraction by population and time (not yet implemented).
}
}
\examples{

# get a data set to work with
X <- slim_sim_dataset_example

tweaked <- tweak_slim_simmed_data(
  X,
  NV = 600,
  ND = 1000,
  S = 0.7,
  RV = 0.9,
  RD = 0.85,
  EV = 0.01,
  ED = 0.005
)


}
