% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wct_rbt_yct_diagnostic_markers}
\alias{wct_rbt_yct_diagnostic_markers}
\title{Diagnostic markers by name}
\format{
A tibble
}
\usage{
wct_rbt_yct_diagnostic_markers
}
\description{
These are WCT, RBT, and YCT diagnostic markers.  Identical to
\code{example_spp_diag_markers} except that the diag_spp column holds
letter codes for the subspecies.
}
\keyword{datasets}
