% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_vcf.R
\name{write_input_vcf}
\alias{write_input_vcf}
\title{Write a tibble of genotypes into a VCF for input as founder genotypes in SLiM}
\usage{
write_input_vcf(Genos, out_path, chrom = "Chr1")
}
\arguments{
\item{Genos}{A tibble with a column POS and then multiple columns
of forward-slash separated genotypes}

\item{out_path}{path to write out the resulting VCF}

\item{chrom}{Name of chromosome to use.  Defaults to Chr1}
}
\description{
Basic and utilitarian.
}
