% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestral_segs.R
\name{ancestral_segs}
\alias{ancestral_segs}
\title{return a tibble with ancestral population of different segments}
\usage{
ancestral_segs(ts, focal_nodes_list, founder_nodes, nodes_tib, indiv_tib)
}
\description{
Uses link_ancestors from tskit, internally using python, to return a tibble
with everything you need to know.
}
