% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_filter_trees.R
\name{read_and_filter_trees}
\alias{read_and_filter_trees}
\title{Read a tree sequence and retain nodes from certain populations at certain times}
\usage{
read_and_filter_trees(
  trees_path = "slim_test.trees",
  years_list = list(`3` = 0:9)
)
}
\arguments{
\item{trees_path}{The path to the .trees file that should be read.}

\item{years_list}{A named list.  Each name is an integer signifying a SLiM
population, and the value is an integer vector of tree-seq years for which
indivdividuals/nodes in the tree seq from the named population should be
retained in the tree-sequence.}
}
\description{
This uses glue to write values into a python script which is then run
by \code{reticulate::py_run_string} which deposits items into the \code{py} variable
in the global environment.
}
